requirements.txt
Django==3.2.12
djangorestframework==3.12.4
psycopg2-binary==2.9.3
requests==2.25.1
Faker==8.12.1
langgraph==0.0.40